requirements.txt
absl-py==2.1.0
accelerate==1.6.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.2
basicsr==1.4.2
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
black==25.1.0
blinker==1.8.2
boto3==1.35.5
botocore==1.35.5
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
config==0.5.1
contourpy==1.2.1
cpm-kernels==1.0.11
cycler==0.12.1
datasets==2.21.0
decorator==4.4.2
decord==0.6.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
einops==0.8.0
etils==1.12.2
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.11
ffmpy==0.4.0
filelock==3.15.4
filterpy==1.4.5
flatbuffers==25.1.24
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
gradio==5.21.0
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.70.0
gviz-api==1.10.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
icecream==2.1.4
identify==2.6.9
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.0.0
isort==6.0.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpickle==4.0.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
lightning-utilities==0.12.0
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.5.5
networks==0.3.7
networkx==3.3
ninja==1.11.1.1
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.20.1
openai==0.28.1
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
peft==0.13.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.2.0
proglog==0.1.10
protobuf==4.25.6
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.20.0
ruff==0.10.0
s3transfer==0.10.2
safehttpx==0.1.6
safetensors==0.4.4
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
spandrel==0.4.1
starlette==0.46.1
streamlit==1.38.0
SwissArmyTransformer==0.4.12
sympy==1.13.1
tb-nightly==2.19.0a20250217
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.19.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tifffile==2025.1.10
timm==0.9.6
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torch-fidelity==0.3.0
torchgeometry==0.1.2
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.0
triton==3.1.0
typed-argument-parser==1.10.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.117
ultralytics-thop==2.0.14
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.30.0
wandb==0.19.1
watchdog==4.0.2
wcwidth==0.2.13
webdataset==0.2.96
websockets==12.0
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.9.4
zipp==3.20.0
zmq==0.0.0
